Refactor TacticalActUsageService - split to separated handler by target type
